requirements.txt
alembic==1.11.1
argcomplete==3.0.8
black==23.3.0
blinker==1.6.2
boto3==1.26.142
botocore==1.29.142
certifi==2023.5.7
cffi==1.15.1
cfn-flip==1.3.0
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.6
cryptography==40.0.2
durationpy==0.5
flake8==6.0.0
Flask==2.3.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
hjson==3.1.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kappa==0.6.0
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
placebo==0.9.0
platformdirs==3.5.1
psycopg2-binary==2.8.4
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
PyYAML==6.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.15
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
troposphere==4.3.2
typing_extensions==4.6.2
urllib3==1.26.16
Werkzeug==2.3.4
zappa==0.57.0
zipp==3.15.0
